requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
langchain==0.0.199
sentence-transformers==2.2.2
pinecone-client
torch
transformers
python-multipart==0.0.5
pydantic==1.10.8
beautifulsoup4==4.10.0
markdown==3.3.6
pypdf2==1.26.0
python-dotenv # Added specific version for dotenv
accelerate>=0.26.0
markdownify 
chardet
beautifulsoup4

